fix screen overflow
