+++ lib/src/pages/home/sub_screens/normal_home.dart
@@ -13,6 +13,7 @@ import 'package:mawaqit/src/themes/UIShadows.dart';
 import 'package:provider/provider.dart';
 
 import '../widgets/FlashWidget.dart';
+import '../widgets/MosqueInformationWidget.dart';
 import '../widgets/footer.dart';
 
 class NormalHomeSubScreen extends StatelessWidget {
@@ -28,31 +29,36 @@ class NormalHomeSubScreen extends StatelessWidget {
       children: [
         Column(
           children: [
-            MosqueHeader(mosque: mosque),
+            Directionality(
+                textDirection: TextDirection.ltr,
+                child: MosqueHeader(mosque: mosque)),
             Padding(
               padding: EdgeInsets.only(top: 4),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Padding(
-                    padding: EdgeInsets.only(left: 2.vw, top: 1.5.vh),
-                    child: ShurukWidget(),
-                  ),
-                  HomeTimeWidget(),
-                  Padding(
-                    padding: EdgeInsets.only(right: 5.7.vw, top: 1.3.vh),
-                    child: Center(
-                      child: SalahItemWidget(
-                        title: S.of(context).jumua,
-                        time: mosqueProvider.times!.jumua ?? "",
-                        iqama: mosqueProvider.times!.jumua2,
-                        active: mosqueProvider.nextSalahIndex() == 2 &&
-                            mosqueProvider.mosqueDate().weekday == DateTime.friday,
-                        removeBackground: true,
+              child: Directionality(
+                textDirection: TextDirection.ltr,
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Padding(
+                      padding: EdgeInsets.only(left: 2.vw, top: 1.5.vh),
+                      child: ShurukWidget(),
+                    ),
+                    HomeTimeWidget(),
+                    Padding(
+                      padding: EdgeInsets.only(right: 5.7.vw, top: 1.3.vh),
+                      child: Center(
+                        child: SalahItemWidget(
+                          title: S.of(context).jumua,
+                          time: mosqueProvider.times!.jumua ?? "",
+                          iqama: mosqueProvider.times!.jumua2,
+                          active: mosqueProvider.nextSalahIndex() == 2 &&
+                              mosqueProvider.mosqueDate().weekday == DateTime.friday,
+                          removeBackground: true,
+                        ),
                       ),
                     ),
-                  ),
-                ],
+                  ],
+                ),
               ),
             ),
             Padding(
@@ -64,69 +70,9 @@ class NormalHomeSubScreen extends StatelessWidget {
             ),
           ],
         ),
-        mosqueConfig!.footer!
-            ? Stack(
-                clipBehavior: Clip.none,
-                alignment: Alignment.bottomCenter,
-                children: [
-                  MosqueInformationWidget(),
-                  Container(
-                    padding: EdgeInsets.symmetric(horizontal: .5.vw, vertical: .2.vw),
-                    width: double.infinity,
-                    color: mosque.flash?.content.isEmpty != false ? null : Colors.black38,
-                    child: SizedBox(
-                      height: 5.vw,
-                      child: FlashWidget(),
-                    ),
-                  ),
-                  Align(
-                    alignment: Alignment.bottomLeft,
-                    child: Padding(
-                      padding: EdgeInsets.symmetric(horizontal: .4.vw, vertical: .2.vw),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            "ID ${mosque.id}",
-                            style: TextStyle(
-                              fontSize: .7.vw,
-                              color: Colors.grey,
-                              fontWeight: FontWeight.bold,
-                              shadows: kHomeTextShadow,
-                            ),
-                          ),
-                          SizedBox(height: 5),
-                          Image.network(
-                            'https://mawaqit.net/static/images/store-qrcode.png?4.89.2',
-                            width: 5.vw,
-                            height: 5.vw,
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(3.0),
-                    child: Align(
-                      alignment: Alignment.bottomRight,
-                      child: Container(
-                          padding: EdgeInsets.only(
-                            bottom: 1.vh,
-                            left: .5.vw,
-                            right: .5.vw,
-                            top: .6.vh,
-                          ),
-                          height: 9.5.vh,
-
-                          child: HomeLogoVersion()),
-                    ),
-                  ),
-                ],
-              )
-            : SizedBox(
-                height: 5.vw,
-              )
+        Directionality(
+            textDirection: TextDirection.ltr,
+            child: Footer()),
       ],
     );
   }
